[FIX] SignUpActivity 코드 정리
