+++ app/src/main/java/com/caffeine/capin/presentation/login/LoginActivity.kt
@@ -18,6 +18,7 @@ import com.caffeine.capin.data.dto.login.RequestLoginData
 import com.caffeine.capin.data.dto.login.ResponseLoginData
 import com.caffeine.capin.data.network.ServiceCreator
 import com.caffeine.capin.data.local.UserPreferenceManager
+import com.caffeine.capin.presentation.signup.SignUpActivity
 import com.caffeine.capin.util.hideKeyBoard
 import dagger.hilt.android.AndroidEntryPoint
 import retrofit2.Call
